Morning Challenge 6
Incorporated Firebase for storing and fetching data.
